Upgrades to the Controller Manager
Check for controllers every 10 seconds, so you don't have to restart the
program just to connect your controllers. I couldn't check every part of
this update (because I can't control when SteamVR assigns left and right
hand roles) but it should work correctly. It definitely works correctly
when unassigned controllers are detected, and when no controllers are
detected, at least in my testing.