Backup/Lots of changes
- NOTE: This update is still a Work In Progress and is not ready for
release. VLCController is not finished, TwitchChat is not implemented,
just included. Mostly pushing this update for code backup. This code has
not been optimized or cleaned and needs work before releasing.

- Hide overlay when selected application closes.
- Add preliminary support for Mouse Clicking (from VR controllers).
Supports clicking but not Click+Drag right now. Seems to be able to
click on every window except MSPaint..
- Added option to Lock Window Size. If the Window's size changes while
it is locked, it will automatically be resized to match the selected
size.
- Upgraded to RenderTexture to cleanly implement:
- Added an Outline to the overlay that currently fades in when you aim
at it, and fades out when you aim away, but will fade in when your
controllers get close enough to "grab" the overlay as well, for
moving/rotating/resizing from VR (not yet implemented)
- Add preliminary support for Multi-Mode. Still needs refining before
release.
- Add preliminary support for additional Capture modes
- Add extra comments, Fix some small bugs, Fix several typos
- Changed the scale such that 1 Unity Unit is now 1 Pixel. This
shouldn't be noticeable but is worth noting (numbers extremely far from
0 can cause issues, so Unity defaults to 100 pixels per unit, but makes
calculations trickier and has no effect when we stay close to 0, so this
shouldn't affect anything at all)
- Probably additional unnoted changes, will try to get a complete change
log finished for the next push.