Updates (mostly fixing mouse interaction)
Note: This version still not completely ready for release. Would like to
add Click+Drag if possible. Still in need of optimization and cleanup,
probably still a few rogue bugs in this that need to be found and fixed
before the next release.
- Changes to the way FPS is selected and handled.
- Device manager no longer spams the log file when only one controller
is connected.
- Adds support for double clicking.
- No longer sends click when it has been too long since you activated
the trigger before releasing it. Click is sent when trigger releases not
when pressed, to match the behaviour of Windows OS. If you press, hold
for a second, and release, it will no longer register as a "click" since
that should be a "click and drag" (not currently supported).
- Fixes clicking with "Send Clicks Only" interaction mode.
- Cursor sprite now fades in and out with Outline.
- Cursor now more responsive.
- Fixes Cursor position when cropping the window.
- Interaction Mode now actually changes things as intended. It is now
possible to click windows in the background as well as clicking without
moving the cursor.
- Save File upgrade to Version 3 for FPS Mode compatibility.
- FPS Mode and Interaction Mode now correctly saved and loaded per
application.
- Outline now starts with 0 Alpha by default.
- Outline minimum now slightly slimmer (Outline thickness is based on
target window resolution, but also has a minimum value, which has been
lowered).
- FPS now defaults to 24, which should be fast enough for any purpose,
AFAIK. Higher FPS rates are probably pointless but I included them
previously in case anyone really needs it.

Not mentioned in the last commit: Changes to the way Capture Mode is
selected. This was noted as "preliminary support for additional capture
modes" but what changed wasn't made clear. Capture Mode is now selected
by a dropdown like Framerate. This is also how Mouse Interaction Mode is
selected, also not noted in the last commit.