Getting closer to v1.0.6 ;]
- Finishes adding Backside support (for now :P)
- Adds "Show Desktop Cursor" support. When enabled the cursor in VR will
mimic the Desktop Cursor unless one of the controllers is interacting
with the Overlay.
- Adds 6 selectable backsides (Green, Blue, Purple, Red, Orange, and
Yellow) as well as "None" to disable the backside.
- Removed Click Interaction Mode and replaced it with more detailed
settings under "Additional Settings".
- Adds Backside to Profile save system and Cursor settings to
Application save system.

I'm pushing the last 4 commits just before making this one. I held off
on pushing them because I've been sick and didn't want to push broken
stuff to the Repo.
The last commits also:
- Added Basic Companion Overlay support, for #30, but I'm not finished
with this yet
- Added ScrollWheel support to certain UI elements, for #31
- Added Quality Settings, for #32, these are under "Additional Settings"

Closes #29