New stuff!
- Animation Panel now changes depending on the Animation being used.
- New Animation Mode: DodgeGaze - DodgeGaze causes the Overlay to move
out of the way when you look at it. End Position and Speed customizable.
- Added Aspect Lock for Target Application Width/Height. Enabling this
causes the Width/Height fields to link and maintain the aspect of the
window when this was enabled.
- Changes for #44. Loading a Profile for the Left controller should now
automatically select the Right Controller when the Left is absent, and
vice versa. Still needs to be thoroughly tested.
- Fixed ScrollWheel on InputFields having rounding issues.
- Updated ScrollWheel InputField script to allow two decimal places on
certain input fields.
- New Companion Overlay Mode: DodgeOnGaze - Used for DodgeGaze
Animation;
- Updates to Companion Overlays to allow Controller attachment,
disconnecting from the Overlay, and fixing RenderTextures/MovieTextures.
- Changes to TrackedDeviceManger to fix Gaze Detection involving
multiple Overlays/CompanionOverlays. Also adds support for
CompanionOverlay Gaze Detection. Still needs thorough testing.
- Changes to RenderTexture script for Backside/GazeDodge Companions that
make them take a single picture through the camera and then disable it
to reduce resource usage. It is reenabled as needed.
- Changes to RenderTexture script for the Primary Overlay so that
resizing windows causes the Overlay to instantly resize (instead of
waiting for the Window to stop Resizing).

Does more work towards #21 , #30, Closes #44