Adds Scrollwheel Support, Fixes DodgeGaze, Adds Config Backup
- DodgeGaze has now been tested on an actual Vive and everything seems
good. Disabled Grabbing while Dodging for now - Just look away and grab
it. I think this is the last one for #21 so closing.
- ScrollWheel Support added for #27. Right now you have to aim and hold
the grips, then you can move your thumb along the touchpad to scroll
horizontally or vertically, but I'm adding the options to disable the
ScrollWheel or disable the Grips requirement. Works on all ClickAPIs as
far as I can tell. Still tweaking the exact bounds for the gesture but
it should be good enough right now; it's a little tricky to get the
bounds just right so that it both feels correct and also supports both
the Vertical and Horizontal Scroll Wheels while also selecting the
correct axis.
- Adds Config Backup for #33. The Application now tests your savefile
after saving and if it's good to go it writes a backup also. If the
savefile is invalid it loads the backup. This _should_ fix the lost
configs issue but you still might lose some settings when the savefile
derps (anything since the last save).

Still gotta make some changes but everything seems functional right now.
This one is pretty close to v1.0.6, though I'd like to add at least one
additional Companion Overlay before release. Also still looking into
Click+Drag.

Closes #21, closes #33